+++ src/index.ts
@@ -29,7 +29,7 @@ if (!TESTRAIL_INSTANCE_URL || !TESTRAIL_USERNAME || !TESTRAIL_API_KEY) {
 
 const server = new McpServer({
     name: "TestRail MCP Server",
-    version: "1.2.0",
+    version: "1.3.0",
 });
 
 const client = new TestRailClient(TESTRAIL_INSTANCE_URL, TESTRAIL_USERNAME, TESTRAIL_API_KEY);